Add changes to try to support iGrill.
Sadly, with the level of BLE support the iGrill
won't handshake with the Particle. It's seen and
added, but doing the authentication as seen
in the Python implementations causes radio
silence. We never get the device challenge data
so can't continue in the process.

I'm wondering if this has to do with the note in
the [python here](https://github.com/bjoernhoefer/igrill/blob/master/igrill.py#L35) that bonding is required.